move page tests into separate file
