+++ fugl/main/tests/models/test_models.py
@@ -5,52 +5,11 @@ Tests for the few model functions we have.
 from django.utils import timezone
 
 from main.models import Category
-from main.models import Page
 from main.models import Post
 from main.models import Project
 
+from .helpers import parse_markdown
 from ..base import FuglTestCase
-
-
-def parse_markdown(md):
-    idx = md.find('\n\n')
-    prefix = md[:idx]
-    suffix = md[idx+2:]
-    pairs = [l.split(':', maxsplit=1) for l in prefix.splitlines()]
-    pairs = [(fst.strip(), snd.strip()) for fst, snd in pairs]
-    frontmatter = dict(pairs)
-    return frontmatter, suffix[:-1]
-
-
-class PageTestCase(FuglTestCase):
-
-    def setUp(self):
-        self.setUpTheme()
-
-        self.project = Project.objects.create(title='project',
-                                              description='project',
-                                              owner=self.admin_user,
-                                              theme=self.default_theme)
-        self.project.save()
-
-        self.page = Page.objects.create(title='page1',
-                                        content='page1',
-                                        project=self.project)
-        self.page.save()
-
-    def tearDown(self):
-        self.page.delete()
-        self.project.delete()
-        self.tearDownTheme()
-
-    def test_filename(self):
-        self.assertEqual(self.page.filename, 'page1')
-
-    def test_get_markdown(self):
-        md = self.page.get_markdown()
-        frontmatter, content = parse_markdown(md)
-        self.assertEqual(frontmatter.get('Title', ''), self.page.title)
-        self.assertEqual(content, self.page.content)
 
 
 class PostTestCase(FuglTestCase):
